Fixed the option to build and launch the local test site to run the unit tests on, since it was apparently broken. Instructions for running the unit tests:
  * mvn clean install
  * mvn site
  * mvn site:run
  * Open the web browser with which you wish to test Sarissa and navigate to http://localhost:8080/test/testsarissa.html
  * Click on the button "Run Testcases".
  * Repeat for every other browser that you want to test Sarissa with.

NOTE: at the time of this writing, 5 out of 8 tests in XSLTProcessorTestCase are failing in Microsoft's new Edge browser that comes with Windows 10. Apparently, that is a known issue with Edge, so Sarissa is probably not to blame for that. Also, all other tests pass in Edge without problems. Other browsers I've tested with so far are IE9, IE11, Safari 5 for Windows and recent versions of Firefox and Chrome. All tests seem to pass in all of them. For more details about the XSLT Processor issue in Microsoft Edge, see https://connect.microsoft.com/IE/feedback/details/2030537/edge-using-javascript-and-xsltprocessor-edge-fails-to-run-transformation-of-stylesheet-having-a-doctype-node