Update LICENSE.md
Corrected the licenses: I had inadvertently appended the wrong licnses to the LICENSE document: Apache 2.0 was correct, but GPL and GPL were supposed to be version **2** and higher, as opposed to version 3 and higher. Also, the license texts are now listed in the same order in which they were specified on the original project website.